+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add iterate view support
 * Improve performance by removing unnecessary dataset traversal
 * Re-add custom config file support
+* Re-add max number of pages support
 
 ### 0.1.0
 * Replace 'c' variable in couchtato.js tasks module with a more descriptive 'util'